Typechecker: migrate to relevant-ish promises by making Promise R/W
This is a relatively experimental change, but changes Promise to be
read/write as an approach to introducing multiple writes to Promise*.
This approach is the simplest way to allow this behavior without
breaking any of the guarantees of Promise* or significantly complicating
the language specification. The typechecker still demotes Promise* -> Promise
upon a write or use as Promise*, but no longer blocks the write operation
to a normal Promise. Another simple alternative here is not demoting the
Promise* automatically, which will allow us to substitute later Promise*s
for an already used Promise. However, this introduces some confusion (in
my opinion) because a Promise* in a function body won't be the same as a
Promise* in a function header (if a function takes a Promise* then we know it
must be used, but if a Promise* exists as a value it won't necessarily have
to be used). As a simple example of where such a solution could be wonky:

func main(): Unit begin
  let u: Promise*(Int) = promise Int in
  u <- 5;
  let v: Promise*(Int) = u in
  ()
end

This function would fail to typecheck after redefining v, because it would
introduce a new linear variable `v` that must be eliminated. However, no
new promise is ever created, so it doesn't make sense to require an extra
write just to get the code to compile.

A few other solutions are possible:

* Promise*(Int) => Promise(Int)
* PromiseRW*(Int) => PromiseRW(Int)

though this gets confusing because of the overloaded use of the *.

Alternatively,

* Promise*(Int) => Promise(Int)
* Promise+(Int) => Promise~(Int)

is a more palatable possibility in my opinion, but significantly expands
the complexity of the language.

One final option that I think is the most elegant would be introducing a
second write operator that only writes to Promise*. This would be useful
because it adds the least number of new features while allowing both
semantics to exist side-by-side.

In addition, there is the question of polymorphism over the *. It seems
like standard library functions could gain value by being overloaded over
both Promise and Promise* while exposing slightly different behaviors. As
such, we want some way to be generic over the input promise type while still
allowing us to choose the write safety mode.