Lang: Add `<~` operator for `unsafeWrite`
This commit adds the <~ operator and reverts the <- operator to its
old behavior, where <- can only write to a Promise*. The motivation
for this is emulating relevant vs. linear promises, where the user
can choose to use a Promise* as if it were either one of these.
Specifically, the <- operator treats the Promise* as linear (once
it has been written to, you cannot write again) and the <~ operator
treats the Promise* as relevant (you can continue writing).

This is one of the possible directions we can take for allowing both
types of substructural "types" to exist in the language simultaneously,
but may change in the future since it has some limitations.

Specifically, the <~ type works on ALL Promises, so it can be a little
confusing when a user writes to a Promise they don't own. Down the line
it may be worth investigating alternatives, so that <~ can only write
to a promise that WAS a Promise*.