Driver: Replace program.programName with a valid Java identifier
When generating the final Java code, it's useful to have a class
and file name that's compatible with Java. To generate this, we
can take advantage of the input filename and strip it down. In
particular, we can remove the file's path and extension and then
filter out all invalid characters. We also capitalize the first
letter to make it a little more idiomatic. As an example:

src/core/core-string.rp

would yield the programName "Corestring".