Driver: Output generated Java code to file
Upon finishing compilation, we use the program.programName value to
create a file called {programName}.java in the current directory
and then write the generated Java code into that file. In the future
there should be a way to select the target directory or even name
the output file, but for now this is a reasonable default.