Parser/AST: Add algebraic type declaration syntax
This commit adds syntax (no codegen/typechecking) for user-defined
algebraic data types. We add two new syntactic constructs:

union NameOfUnion begin
  Case1;
  Case2(TypeContained)
end

record NameOfRecord begin
  field1: Type1;
  field2: Type2
end

The `union` construct defines a sum type with named cases (similar
to enums in Rust) and the `record` construct defines a product type
with named fields (similar to structs). For now, these types will
always be immutable, which allows us to guarantee that the values
contained are acyclic. The purpose of this restriction is that we
can then prove that we visit each node of a heap-allocated value.

With such a proof, it's possible for us to lift the * type to the
top-level of a container type, so that List(Promise*(Int)) becomes
List*(Promise*(Int)). We can then prove that when each node is
visited, the promise contained is written to. (This is yet to be
implemented).