CG: Implement code generation for user-defined types
This PR implements rudimentary codegen for the user-defined union
and record types. Records are generated as POD (plain old data)
classes where all members are public and feature only a constructor.
Unions, on the other hand, are generated as an abstract base class
with each case having its own POD to extend that base class (which
accomplishes the proper subtyping relationship needed for unions).

Both of these types also make use of a ThreadLockedObject base
class, which will be used as the basis of every heap object in the
language. The goal here is to allow the generated code to verify
that a TLO (thread-locked object) belongs to its private heap. In
turn, this allows us to make sure that no promises can be shared
between multiple owners by being placed in a heap object.

The next step here is to generate code that uses these types.
Specifically, we want the user to be able to
1) create objects from user-defined types
2) access fields from these types
3) match against these types