UserTypes: Add value construction for user-defined types
User types can now be constructed and the syntax has slightly changed.
Union cases now use brackets instead of parens and always require the
brackets. E.g. Empty[] and Node[IntList, Int] instead of Empty and
Node(IntList, Int). This syntax is used both in the constructors and
in the type declaration. Records can be constructed with the
constructor followed by the params enclosed in braces. E.g.:
Pair { first = 1, second = 2 }

This commit also implements type checking, where we ensure that each
applied argument for a constructor is of the right type. For named
parameters in records, we perform reordering of the arguments so that
out-of-order arguments are matched. We also make sure that the correct
number of arguments are always applied.

The generated code is straight-forward, simply calling out to the
proper Java constructor for the type (or case if it's a union). For
records specifically, the reordering is used so that out-of-order
arguments are placed in the correct slot when calling the Java
constructor.