CG: Add .match() method to generated code for union types
This creates a match method on unions for easier pattern matching.
Essentially, the match method takes an anonymous function for each
branch in a match expression. Based on the case class that it's
being called on, the correct function is dispatched. Each class
automatically passes in all of its args to the function, so that
these are already unwrapped in the matching branch.