UserTypes: Add record access expression
Expressions of the type r.x where r is a record and x is a field are
now supported. Type-checking is performed to ensure that the field
actually exists in the type. The generated code is a direct access
to the field in the record class, but this may change in the future
to either use getters or include an access check before actually
attempting the record field access operation.