CG: Add codegen for match expressions
Match expressions are generated as a list of anonymous functions passed into
val.match() in the declared order. For now, we don't generate the type annotations
for the args as these are optional in Java but this may need to be added later.
Additionally, we only generate a single expression which could be an issue in the
future. A quick fix would be surrounding the values here with {} and setting
~in_expr:false but this may lead to some other problems.