Parser: Remove shift/reduce conflicts for question/dot precedence
Previously, the following expression was ambiguous: `?p.q`. It was
unclear whether this meant `(?p).q` or `?(p.q)`, which resulted in
a number of shift/reduce conflicts in the parser. This is fixed by
making the DOT operator have a higher precedence than the question
operator. As a result, the expression is now parsed as `?(p.q)` if
the user doesn't use parens. A test case has been created for this
as question_dot_precedence.txt.