Typecheck: Function parameters should become *-values when liftable
Until now, when a function took a Promise* parameter it was introduced
to the linear environment in the function body. This generalizes that
behavior to all liftable types. For example, if we have a record type
that contains a Promise*, that should also be introduced to the linear
environment in that function.