Runtime: Add ThreadLockedObject
ThreadLockedObject (or TLO) is the base class that all user-types inherit
from. This class allows us to perform an access check in order to make sure
that a value is on the local thread's private heap. The goal here is to
make user-defined types safe to use. This comes down to a few rules, one of
which is that all heap-allocated values must be explicitly owned by their
creator thread, and for now ownership cannot be transferred. If this were
not the case, it would be possible for another thread to fulfill the
promises contained in a heap value without the owner (or type system)
realizing, causing confusing type errors. For now only a dynamic check
exists but hopefully we'll be able to make this work statically in the
future.