Typecheck: Liftable match variables are introduced to linear environment
When we match against a union constructor that contains liftable values,
we want to introduce these to the linear environment so that the user is
forced to consume them. Previously, we only applied this to Promise*'s
contained in constructors, but this has been generalized to introduce
all liftable values. The goal here is to deal with recursive types (such
as a list of promises) where each successive Cons node must be checked
recursively. This, along with the exhaustiveness check in matches,
allows us to guarantee that an entire value (which is a tree) has been
inspected exhaustively.