Typecheck: Check that let body's linear environment is a subset of parent environment
This fixes a regression in unbound-liftable.txt. Previously, we could leak linear vars
by constructing them into a value and then ignoring it. This worked because the let
performed an intersection after adding the new linear variable, so the value would
disappear even if we did nothing with it. To fix this, we simply check that the new
linear environment is a subset of the old one, which is enough to tell us that this
new variable was eliminated (as expected) before we apply the intersection operation.