Lexer: Add syntax for comments
Finally got around to adding comments. This adds two types of comments.
One is inline comments, which start with -- and go until the end of the
line. The other is multiline comments which start with (* and end with
*). Multiline comments can't be nested (for now, this might change if
it ever becomes an issue). A simple test file comments.txt has been
added to test this behavior.