Typecheck: Allow if branches to eliminate different linear vars
Previously, if branches needed to eliminate the same set of linear
variables in order to typecheck. However, this meant that we could
not conditionally write to a promise in one branch and then ignore
it in another. Since we can still prove that a program writes to a
promise by writing after a conditional, this restriction can now be
loosened. Instead, we take the union of the two linear environments
in each branch and use that to calculate changes to the final linear
environment/normal environment. This is tested in branch-elim.txt.

This restriction has not been lifted in matches, as our "move
semantics" cannot handle this case yet. In the future, this can
probably be solved by just making the match a nop (at the type-level)
if there are leftover linear variables.