Typecheck: Move evalArgs to top-level
This is a quick change to enable more code reuse. Previously, we had
two (nearly identical) definitions of evalArgs in our typecheck
function, where one was for function params and the other was for
union construction. Since the only difference was the errors printed
out, this changes the error string to be configured by a parameter that
specifies location (i.e. in a function application or in a union
construction) when creating error messages. Apart from that, the function
has simply been hoisted out of typecheck into a global function.