Typecheck: Perform preprocessing of identifiers to make them unique
Previously, there was a minor bug where we could not safely reuse
variable names, since the environment splitting was not aware of
what scope each variable came from. This adds a really basic support
for this by pre-processing variable names before the program reaches
the type checker. Essentially, each identifier is given a unique ID
and replaced with "identifier%uniqueID". Since a special/unrepresentable
character is used to split the token, we can easily extract the original
name from this string. At the same time, the idents are guaranteed to
be unique. A special pass is inserted to preprocess each function and
replace every ID with a unique new ID, recursively moving down the tree
and keeping track of the current "Env" of replacement idents (i.e. we
keep track of the fact that "k" maps to "k%99", etc.).

This commit also does some minor refactoring in where the Env/Vars
modules live, but this might need some future rethinking.

This solution seems to work ok, since now every
variable will be uniquely named, but the debug outputs could be a
little confusing since custom names are generated to show the scope
each variable came from. To get around this, we can patch env/expr
printing to no longer show the full name, cutting it off at the %
character (which represents the "extra" scope information).