lang: Change syntax for promise instantiation
This change makes it so that promises are declared as a statement
rather than as an expression. I.e. the syntax
```
let ... = promise Int in ...
```
is now
```
promise r, w: Int in ...
```

This change was made due to the differences in product types between
the compiler and the paper -- specifically, the compiler does not
allow for either anonymous or generic product types, so it doesn't
make sense for the `promise` operator to return a product type.
However, with the goal of simplifying the type system this seems
like a pretty big win :)