CG: fix implicit returns in the context of loops
It's kind of a hack to get implicit returns working well in
Java, and until now it was very broken in the case of loops.
For example, given the program:
```
func f(): Unit begin
  for x = 0 to 10 begin
    ()
  end
end
```
the following code was generated:
```
public static Unit f() {
  for (Integer x = 0; x < 10; x++) {
    return Unit.the; // Return where it doesn't make sense
  }
  // Lack of final return fails to compile in Java
}
```

The two issues here were both traced to the same core bug:
we had no way of discerning whether an expression represented
the final expression in a program. By adding in a ~is_final
flag to the emit function, we can keep track of what the final
expressio is; each call to emit passes down the ~is_final flag
to its final expression and sets the others to false. This is
leveraged here in a few ways: firstly, we can be sure not to
emit a return in the body of a loop, since these are known not
to be final; secondly, we know to emit a return at the end of
the function after the loop when the loop is the final expression.

This new behavior _did_ introduce a bug in the case of statements
that do not return, as Java does not allow any arbitrary expression
as a statement. For example, `Unit.the;` is not a valid statement even
when `return Unit.the;` is. To get around this, we can simply use a
`Unit.ignore` function to throw away the result but promote this
expression into a valid statement.