lang: add string literals & string type
This adds basic support for strings to the language, though there
are no operations on strings available yet (outside of comparisons).
Additionally, there may be a bug in code gen for strings: if a string
needs to be escaped, the compiler does not know to escape the string
correctly in the Java way. It does, however, escape it for OCaml
string escape syntax which is probably close enough. I don't intend to
fix that behavior.