typechecker: add no-must-use restriction to loops
Previously, loops could own a linear value. Since a loop executes
an indeterminant number of times, there are two possible bugs:
1. the loop never uses the linear value
2. the loop reuses the linear value

As an example:
```
func loop(times: Int): Unit begin
  promise r, w: Int in
  for i = 0 to times begin
    w <- 5
  end
end
```

For times = 0, we would never write and for times = 2 we
would rewrite on the linear value. With relevant values,
the second case is no longer problematic but the first
case would still be a bug.

The simplest fix here is to simply disallow the ownership of
linear values by loops. On the other hand, this really restricts
the power of loops (which shouldn't be a huge deal, but recursion
is pretty bad in Java so we kinda need loops sometimes).

Ideally, unsafe writes would be a good tool here, but unfortunately
unsafe writes are also in a tricky spot right now -- because of
how environment splitting works, it requires the linear value to
be present in only one environment. This means that once we perform
the write, the variable effectively no longer exists. Beyond this,
even switching to relevant types alone is not a sufficient fix, as
there is no evidence that the value actually gets used in a safe
write by just looking at the environment. On the other hand, I think
it's safe to use the read end of a promise unsafely, so this should
hopefully be good enough.