typechecker: switch unsafe writes to work on read end of promise
This change is kinda funky, but as described in the last commit, it's
really only safe to do unsafe writes on the read end of a promise
right now. Since the promise* "dies" one it's used it obviously
can't be reused in another scope to write, so I think this change is
somewhat appropriate and can be explained to a user intuitively.