lang: add type inference in let expressions
Let expressions now allow for 1-directional type inference. For example,
the program:

let x = 0 in ...

Is valid, because the right-hand side type information is propagated to
the left-hand side. Since we unambiguously know the right-hand type in
all expressions (without the need for any unification), we can easily
implement this style of inference. Note that this may present issues
if we introduce generics later on.

Right now, this is implemented by leaving a place-holder `Infer ...
type in the syntax tree, which contains a `ty option ref`: once a
type has been inferred, we can populate the tree with the updated
info in-place and the code generation, etc. will base its analysis
on this information.

I am not a fan of the fact that the `Infer type is globally possible
and originally wanted to restrict it to only existing in the syntax
for let, but unfortunately that would take a lot of GADT magic to get
working as far as I can tell, so for now this should be good enough.
Just hope that I didn't forget to update any code to handle this case
(i.e. match on all types with an _ that doesn't handle infers).