cg: insert returns in match bodies
When we added the ~is_final flag to the `emit` function, we introduced
a way to track whether a return or an ignore is needed for any given
expression, i.e. only the final expression in a function should be an
implicit return. However, the way code generation for matches works,
we insert a new body `(...) -> { ... }` rather than a single expression
`(...) -> ...`, so we have essentially introduced a new function. As a
result, we were previously not emitting the `return` necessary in a
match body. This led to an interesting situation where a program:
```
func x(y: LeftOrRight): Int begin
  match y begin
    Left[] -> -1
    Right[] -> 1
  end
end
```
would typecheck in our compiler, but fail to typecheck in the Java
compiler since each branch of the match expression would really be
returning `Unit`.

The fix here was simple: we only need to mark the branches of a match
as final expressions.