runtime: Add sleeping capabilities
This adds three functions: sleep(), milliseconds(), and seconds().
The sleep() function allows the current thread to sleep for a specified
period of time, while milliseconds() and seconds() allow us to convert
an integer amount of time into a period understood by sleep(). For example
to sleep for 10ms we can do `sleep(milliseconds(10))` and to sleep for
5s we can do `sleep(seconds(5))`.