examples: add ports of stackoverflow JS programs that fail to compile
Taken from case study in https://dl.acm.org/doi/pdf/10.1145/3133910,
these examples demonstrate dead promises and double resolves that are
found in the wild. Some liberty had to be taken when converting code
since not all JS features map well into our own features, but in general
the behavior of the programs emulates the behavior of the original code.