runtime: Switch Promise to just forward to CompletableFuture
There's no reason to roll our own promises, so we switch to using
the Java built-in implementation. Since we weren't using the
capabilities of ThreadLockedObject (and because objects are always
read-only in our language anyways), we simply drop the dependency.
Because our language doesn't have ref cells, we've been using
promises for synchronous (thread-local w.r.t. writes) state. As a
result, we need to bend CompletableFuture a little bit to allow us
to alter the stored value after it's been completed. Luckily, we
can simply use .obtrudeValue() if the promise .isDone() :)