runtime: Drop ThreadLockedObject
ThreadLockedObject was something we added early on in the runtime but
never actually used (and it wasn't even fully implemented!). Since
we don't actually need its capabilities because of the lack of mutable
state, we can simply drop all references to it in our generated code
and everything will still work!