runtime/CG: shutdown thread pool at end of main() function
Previously, after submitting all the tasks to the thread pool the
generated program would hang forever (until interrupted). This was
the case because Java's ExecutorService class continues to wait for
new tasks to be submitted. To get the program to no longer hang, we
need to tell the ExecutorService to stop waiting for new tasks by
calling the shutdown() method. This would cause the pool to exit
immediately, so we need to also tell it to wait (with an infinite
timeout) until all submitted tasks are done executing. The end
result is a program that shuts down as soon as all async tasks are
done executing.