javaCG: generate intermediate if statement correctly
Previously, the following construct:

   if x then y else z end;
   ()

would cause us to always generate the Java code:

   Unit $_0 = ((x) ? (y) : (z));
   return Unit.the;

While this looks reasonable, consider the case where y is itself
a compound expression:

   if x then a; b else z end;
   ()

This would previously generate the incorrect Java:

   Unit $_0 = ((x) ? (a; b) : (z));
   return Unit.the;

Clearly, the above is not valid Java. The issue here was that we would
always assume that it's safe to generate a Java _expression_ conditional
in a let expression. However, a let _ = ... expression is actually a
special case in which the RHS is unlikely to be a valid expression. To
fix this special case, we can simply override the code gen for any expr

   let _ = if ... then ... else ... end; ...