types: allow for promises of promises/promise stars
This removes the split between primitive types & normal types. This
split was previously used to make it impossible to create a type like
Promise(Promise*(Int)). However, this _is_ a safe construct:

1. Promise(Promise*(Int)) cannot duplicate a Promise*(Int):

   Because Promise(Promise*(Int)) is linear, we cannot duplicate it.
   We can only use it by transferring ownership of the value (via call
   or move) OR with the ? operator. The ? operator uses the promise and
   gives us a new linear value, Promise*(Int) that also cannot be
   duplicated.

2. Promise(Promise*(Int)) cannot drop a Promise*(Int):

   Similar logic applies: We cannot drop the promise, as it is linear.
   Once we read the promise, we are left with a linear variable of type
   Promise*(Int) that also cannot be dropped.

The only possible error that can exist here is in the interaction of
unsafeWrite (<~) and Promise*(Promise*(Int)):

    promise getP, setP : Promise*(Int) in
    promise getA, setA : Int in
    promise getB, setB : Int in
    setP <- setA;
    getP <~ setB -- Now setA is lost -> omitted write!
    -- getP is now also used. So not only is setA lost, but setB is
    -- also lost! This leaves us with _2_ omitted writes!

Given that unsafeWrite is _unsafe_ this is probably a reasonable thing
to allow. However, I am hesitant to enable this because it does break the
soundness of unsafeWrite as described in the paper.