runtime/AsyncTask: use Function0<Unit> to back AsyncTask
Previously, AsyncTask was backed by a Runnable, which is a lambda of
void(). This was semantically incorrect, because async expected a
Unit() lambda. In fact, this could allow us to generate incorrect code
in certain situations:

    async (if x then () else () end)

would generate:

    new AsyncTask(() -> ((x) ? Unit.the : Unit.the))

which would fail to compile in Java. Interestingly, Java allows for:

    new AsyncTask(() -> Unit.the)

which hid the error until now.