CG/runtime: use closures to convert between exprs/statements
Our language is an expression-based language, whereas Java uses both
expressions & statements. For example, in our language a loop is an
expression that yields unit, so we can say something like this:

    let _ = while true begin () end

This has no simple equivalent in Java. However, Java _does_ give us a
way to translate any statement (such as a while loop) into an expression:
closures. We can rewrite the above like so:

    Unit $_0 = ((Function0<Unit>)(() -> {
      while (true) {
        Unit.ignore(Unit.the);
      }
      return Unit.the;
    })).apply();

The compiler was already doing most of the translation here, allowing us
to sequence multiple statements into a valid lambda. However, we were
manually making conversions between exprs & statements in quite a fragile
way (as noted by the previous few commits addressing bugs in code gen that
were caused by expr/stmt translation issues).

However, this solution isn't a silver bullet for a few reasons:

1. We have to split exprs into two classes of exprs, those that can be
compiled to statements & those that can be compiled to expressions. This
property is not currently encoded in the AST and is simply handled by pattern
matching, which leads to an undesirable effect: if we were to add a new type
of expression, we lose exhaustiveness checks in the codegen functions. There's
a better way to do this (using polymorphic variants we can have a type for
both cases), but it isn't a super quick translation so I've held off on it for
now. The second side of this coin is that we've removed the overlap of things
that are both exprs and statements, e.g. conditionals and let _ = [...]. This
isn't the end of the world, but it leads to some subpar code and will take
some fixing (I think it's non-trivial to decide when to use expr/stmt for these).

2. Java syntax is funky about applying lambdas without a type. Note the cast in
the code shown above: this is necessary for Java to compile the program. In
particular, to use such a cast we need to know both the type of the closure and
the type of the return value. These are both things we know at compile-time, but
getting the return value's type is a little harder. Thankfully, Java gives us an
easy way to get around this:

    public class Function0<R> {
      ...
      public static <R> R call(Function0<R> fn) { return fn.apply(); }
    }

This gives us a short-hand way to apply a function & solves this issue!

3. Java's semantics make it impossible to capture a non-final variable in a
closure. For example, we cannot write:

    for (int i = 0; i < 10; i++) {
      Function0.call(() -> { System.out.println(i); });
    }

because i is mutable (even though the closure can't mutate it & it can't be
mutated while the closure is executing). Thankfully, there is only one place
in the language that actually mutates a variable: for loops (note, changing an
object's internal state does not count as mutation in Java, so Promise::fulfill
is safe). This means that we can safely translate all programs, with one minor
change: we only need to make a final copy of the loop variable on each iteration.

    for (int $_0 = 0; $_0 < 10; $_0++) {
      final int i = $_0;
      Function0.call(() -> { System.out.println(i); });
    }

This works because the variable changes only upon each iteration, so making this
copy of it is completely safe!