compiler: allow for reuse of idents in generated programs
Previously, we mapped all IDs to unique strings in the typechecker. This
allowed, e.g.:

    let l = 0 in
    let l = "hello" in
    println(l)

to be transformed to:

    let l$0 = 0 in
    let l$1 = "hello" in
    println(l$1)

allowing proper shadowing. Unfortunately, this was only able to take
effect in the type checker, and the results were thrown away before code
generation. This simple fix brings back this behavior so that generated
programs are able to compile under the Java compiler (which disallows
shadowing).