Dockerfile: clear out ~/.opam before installation to allow broken menhir pkg
See https://github.com/ocaml/opam-repository/pull/18697

The current OPAM docker image is not up to date with the latest opam repo &
does not allow for easily updating the local opam repo. As a result, the
changes in the linked PR are not visible from within the Docker VM, which
means that we cannot successfully install Menhir. To force an OPAM update,
we simply remove the ~/.opam dir and reinit OPAM from scratch. This is a
hack but until the OPAM Docker image is patched we rely on this to build a
working image.